[refactor]: Ladder 인터페이스 제거 및 클래스명 수정, 테스트 코드 수정
