+++ src/main/java/kr/codesquad/Ladder.java
@@ -1,7 +1,18 @@
 package kr.codesquad;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
-public interface Ladder {
-    List<String> createOutputLines();
+public class Ladder {
+    private final List<LadderLine> ladderLines;
+
+    public Ladder(List<LadderLine> ladderLines) {
+        this.ladderLines = ladderLines;
+    }
+
+    public List<String> createOutputLines() {
+        return ladderLines.stream()
+                .map(LadderLine::toString)
+                .collect(Collectors.toList());
+    }
 }
